Addquote and randomquote now work

+++ ConsoleApp/DataModels/Quote.cs
@@ -6,8 +6,8 @@ namespace ConsoleApp.DataModels
 {
     public class Quote
     {
-        public long QuoteId { get; set; }
-        public long ServerUserId { get; set; }
+        public int QuoteId { get; set; }
+        public int ServerUserId { get; set; }
         public string QuoteBody { get; set; }
         
         public virtual ServerUser ServerUser { set; get; }
